✨ First working version
We detect defined alerts and create one sensor for each.
For now we don't use the /api/v1/alerts endpoint which returns only the
firing alerts (and we want all of them).

In theory, prometheus recommends to query the alertmanager instead of
prometheus but I want to keep things simple for now (a personal
deployment of prometheus is likely to be prometheus only rather than the
full setup Prom+AM)

Change-Id: Id8f2a920c239a665da81a934d789ae3168adef17